+++ NEWS.md
@@ -1,10 +1,14 @@
 <div style="text-align: left;">
 
-### brmss 0.3 (2025-10-30) (GitHub Only)
+### brmss 0.3 (2025-10-31) (GitHub Only)
 
-* Fix minor bugs in `BVS_subfunc::gammaMC3Proposal()` and `BVS_gaussian::sampleGamma()`
+* Fix minor bugs in `BVS_subfunc::gammaMC3Proposal()` and `BVS_gaussian::sampleGamma()
+* Fix bug with Gibbs sampling for Bernoulli probability
+* Add hierarchical related regressions (HRR)
 
+* TODO: define N, L, p as public (global) variables in each family class
 * TODO: Check if better techniques instead of hard upper- or lower-bounds for exponential/log scales!
+* TODO: add mvprobit
 * TODO: add GLM for gaussian, poisson, bernoulli (logit, logistic, clog-log), gamma regression
 * TODO: add logistic regression with Pólya-gamma augmentation
 * TODO: add multivariate linear regression (mgaussian) [No! BayesSUR is already good for mgaussian.]